feat: reorganize backend submodule to include particle stack functions
