+++ src/tt2dtm/backend/__init__.py
@@ -1,0 +1,10 @@
+"""Submodule for computationally intensive backend functions."""
+
+from .core_match_template import core_match_template
+from .core_refine_template import core_refine_template, cross_correlate_particle_stack
+
+__all__ = [
+    "core_match_template",
+    "core_refine_template",
+    "cross_correlate_particle_stack",
+]
